Enhance petscnagfor to support -Xlinker and @filewithlistofobjects
Needed by new --download-fblaslapack that uses the standard distribution of LAPACK that uses CMake

Note that the standard distribution of LAPACK CMake has multiple bugs in it.

It includes add_subdirectory(LAPACKE) unconditionally instead of only if LAPACKE is on (see the similar construct for CBLAS in the same file). This needs to be fixed by the LAPACK team

This commit should resolve the arch-ci-linux-nagfor CI problems

Commit-type: portability-fix
/spend 5h